refactor: streamline LLMEvaluatorConfigForm by removing evaluator prop and initializing state directly in EditPanel
