+++ web/components/shared/CreateNewEvaluator/LLMEvaluatorConfigForm.tsx
@@ -54,41 +54,7 @@ export type LLMEvaluatorConfigFormPreset = {
   testInput?: TestInput;
 };
 
-const getInitialState = (
-  evaluator?: Evaluator
-): LLMEvaluatorConfigFormPreset => {
-  if (!evaluator) {
-    return LLM_AS_A_JUDGE_OPTIONS[0].preset;
-  }
-
-  const template = evaluator.llm_template as any;
-  const property =
-    template.tools[0].function.parameters.properties[evaluator.name];
-  const propertyDescription = property?.description;
-  const rangeMin = property?.minimum;
-  const rangeMax = property?.maximum;
-
-  return {
-    name: evaluator.name,
-    description: propertyDescription,
-    expectedValueType: evaluator.scoring_type
-      .replace("LLM-", "")
-      .toLowerCase() as "boolean" | "choice" | "range",
-    includedVariables: {
-      inputs: true,
-      promptTemplate: false,
-      inputBody: false,
-      outputBody: true,
-    },
-    choiceScores: evaluator.scoring_type === "LLM-CHOICE" ? [] : undefined,
-    rangeMin: evaluator.scoring_type === "LLM-RANGE" ? rangeMin : undefined,
-    rangeMax: evaluator.scoring_type === "LLM-RANGE" ? rangeMax : undefined,
-    model: template.model,
-    testInput: undefined,
-  };
-};
-
-const useLLMConfigStore = create<{
+export const useLLMConfigStore = create<{
   LLMEvaluatorConfigFormPreset: LLMEvaluatorConfigFormPreset;
   setLLMEvaluatorConfigFormPreset: Dispatch<
     SetStateAction<LLMEvaluatorConfigFormPreset>
@@ -97,7 +63,7 @@ const useLLMConfigStore = create<{
   devtools(
     persist(
       (set) => ({
-        LLMEvaluatorConfigFormPreset: getInitialState(),
+        LLMEvaluatorConfigFormPreset: LLM_AS_A_JUDGE_OPTIONS[0].preset,
         setLLMEvaluatorConfigFormPreset: (by) => {
           if (typeof by === "function") {
             set((state) => ({
@@ -120,11 +86,10 @@ const useLLMConfigStore = create<{
 );
 
 export const LLMEvaluatorConfigForm: React.FC<{
-  evaluator: Evaluator;
   onSubmit: () => void;
   existingEvaluatorId?: string;
   openTestPanel?: () => void;
-}> = ({ evaluator, existingEvaluatorId, onSubmit, openTestPanel }) => {
+}> = ({ existingEvaluatorId, onSubmit, openTestPanel }) => {
   const notification = useNotification();
 
   const jawn = useJawnClient();
@@ -172,12 +137,6 @@ export const LLMEvaluatorConfigForm: React.FC<{
     configFormParams.testInput,
     setTestData,
   ]);
-
-  useEffect(() => {
-    if (evaluator) {
-      setConfigFormParams(getInitialState(evaluator));
-    }
-  }, [evaluator, setConfigFormParams]);
 
   return (
     <Col className="h-full flex flex-col">
